requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
WTForms==3.1.2
supabase==2.3.5
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1